Safari-611.1.10/LayoutTests/media/track/video-track-expected.txt


In-band video tracks.

EVENT(addtrack)
EXPECTED (event.track == 'video.videoTracks[0]') OK

EVENT(canplaythrough)

** Check initial in-band track states
EXPECTED (video.videoTracks.length == '1') OK

** Unload video file, check track count
RUN(video.src = '')
EXPECTED (video.videoTracks.length == '0') OK

END OF TEST
